Fixing build machine error in MediaSoundComponent.cpp #jira none #rb none #robomerge ignore
[CL 7941728 by Thomas Engel in 4.23 branch]

+++ Engine/Source/Runtime/MediaAssets/Private/Assets/MediaSoundComponent.cpp
@@ -422,9 +422,9 @@ int32 UMediaSoundComponent::OnGenerateAudio(float* OutAudio, int32 NumSamples)
 		}
 		else
 		{
-			const int32 FramesRequested = NumSamples / NumChannels;
+			const uint32 FramesRequested = uint32(NumSamples / NumChannels);
 			uint32 JumpFrame = MAX_uint32;
-			uint32 FramesWritten = Resampler->Generate(OutAudio, OutTime, (uint32)FramesRequested, Rate, Time, *PinnedSampleQueue, JumpFrame);
+			uint32 FramesWritten = Resampler->Generate(OutAudio, OutTime, FramesRequested, Rate, Time, *PinnedSampleQueue, JumpFrame);
 			if (FramesWritten == 0)
 			{
 				return 0; // no samples available
